fix: shared object being processed two times causing db errors
