+++ src/hooks/useApp.tsx
@@ -4,21 +4,17 @@ import { useEffect } from "react";
 import { lastAction, displayConfirmation, openDevMode, languageSelectionState, displayToast } from "@src/store";
 import { getTheme } from "@src/store/ThemeState";
 import { checkMagicGoal, getAllLevelGoalsOfId, getGoal, updateSharedStatusOfGoal } from "@src/api/GoalsAPI";
-import { addSharedWMGoal, getSharedWMGoal, updateSharedWMGoal } from "@src/api/SharedWMAPI";
 import { createDefaultGoals } from "@src/controllers/NewUserController";
 import { refreshTaskCollection } from "@src/api/TasksAPI";
-import { checkAndUpdateGoalNewUpdatesStatus, handleIncomingChanges } from "@src/helpers/InboxProcessor";
-import { getContactSharedGoals, shareGoalWithContact } from "@src/services/contact.service";
-import { updateAllUnacceptedContacts, getContactByRelId, clearTheQueue } from "@src/api/ContactsAPI";
+import { checkAndUpdateGoalNewUpdatesStatus } from "@src/helpers/InboxProcessor";
+import { shareGoalWithContact } from "@src/services/contact.service";
+import { updateAllUnacceptedContacts, clearTheQueue } from "@src/api/ContactsAPI";
 import { useSetRecoilState, useRecoilValue, useRecoilState } from "recoil";
 import { scheduledHintCalls } from "@src/api/HintsAPI/ScheduledHintCall";
 import { LocalStorageKeys } from "@src/constants/localStorageKeys";
 import { checkAndCleanupTrash } from "@src/api/TrashAPI";
-import { SharedGoalMessage } from "@src/Interfaces/IContactMessages";
 import { getAllInboxItems } from "@src/api/InboxAPI";
-import { Payload } from "@src/models/InboxItem";
-import { useQuery } from "react-query";
-import { GoalActions, TaskActions } from "@src/constants/actions";
+import { TaskActions } from "@src/constants/actions";
 
 const langFromStorage = localStorage.getItem(LocalStorageKeys.LANGUAGE)?.slice(1, -1);
 const exceptionRoutes = ["/", "/invest", "/feedback", "/donate"];
@@ -33,44 +29,6 @@ function useApp() {
 
   const confirmationState = useRecoilValue(displayConfirmation);
 
-  const handleNewIncomingGoal = async (ele: SharedGoalMessage, relId: string) => {
-    const { levelGoalsNode } = ele;
-
-    try {
-      await Promise.all(
-        levelGoalsNode.map(async (goalNode, index) => {
-          const { goals } = goalNode;
-          await Promise.all(
-            goals.map(async (goal) => {
-              try {
-                const existingGoal = await getSharedWMGoal(goal.id);
-                if (existingGoal) {
-                  const { parentGoalId } = goal;
-                  if (parentGoalId === "root") {
-                    console.log("No parent goal id");
-                    return;
-                  }
-                  await updateSharedWMGoal(goal.id, {
-                    ...existingGoal,
-                    parentGoalId,
-                  });
-                } else {
-                  await addSharedWMGoal(goal, relId);
-                }
-              } catch (error) {
-                console.error("[handleNewIncomingGoal] Error processing goal:", error);
-              }
-            }),
-          );
-        }),
-      );
-
-      setLastAction(GoalActions.GOAL_NEW_UPDATES);
-    } catch (error) {
-      console.error("[useApp] Error processing shared goals:", error);
-    }
-  };
-
   useEffect(() => {
     getAllInboxItems().then((inboxItems) => {
       inboxItems.forEach((inboxItem) => {
@@ -80,24 +38,6 @@ function useApp() {
       });
     });
   }, []);
-
-  const { data: sharedGoalsData } = useQuery({
-    queryKey: ["contactSharedGoals"],
-    refetchOnWindowFocus: false,
-    queryFn: async () => {
-      const res = await getContactSharedGoals();
-      if (!res.success) {
-        throw new Error("Failed to fetch shared goals");
-      }
-      return res.response.reduce(
-        (acc: { [key: string]: SharedGoalMessage[] }, curr) => ({
-          ...acc,
-          [curr.relId]: [...(acc[curr.relId] || []), curr],
-        }),
-        {},
-      );
-    },
-  });
 
   useEffect(() => {
     const init = async () => {
@@ -146,34 +86,6 @@ function useApp() {
           }),
         );
       });
-
-      if (sharedGoalsData) {
-        // Process all relIds in parallel using Promise.all
-        await Promise.all(
-          Object.keys(sharedGoalsData).map(async (relId) => {
-            const contactItem = await getContactByRelId(relId);
-            if (contactItem) {
-              // Process all changes for this contact in parallel
-              const changes = sharedGoalsData[relId];
-              await Promise.all(
-                changes.map(async (change) => {
-                  try {
-                    if (change.type === "shareMessage") {
-                      console.log("change", change);
-                      await handleNewIncomingGoal(change, relId);
-                    } else if (["sharer", "suggestion"].includes(change.type)) {
-                      await handleIncomingChanges(change as unknown as Payload, relId);
-                      setLastAction(GoalActions.GOAL_NEW_UPDATES);
-                    }
-                  } catch (error) {
-                    console.error("Error processing change:", error);
-                  }
-                }),
-              );
-            }
-          }),
-        );
-      }
     };
     const installId = localStorage.getItem(LocalStorageKeys.INSTALL_ID);
     if (!installId) {
@@ -187,7 +99,7 @@ function useApp() {
     if (!isLanguageChosen && !exceptionRoutes.includes(currentPath)) {
       window.open("/", "_self");
     }
-  }, [langFromStorage, sharedGoalsData]);
+  }, [langFromStorage]);
 
   useEffect(() => {
     localStorage.setItem(LocalStorageKeys.CONFIRMATION_STATE, JSON.stringify(confirmationState));
